backend/requirements.txt
bleach==6.0.0
crispy-tailwind==0.5
djlint==1.19.15
django-money==3.0.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==9.4.0
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
whitenoise==6.3.0
psycopg2-binary==2.9.5
